Last Update

+++ Assets/Scripts/EnemyScripts/EnemySO.cs
@@ -6,14 +6,14 @@ public class EnemySO : ScriptableObject
     public GameObject EnemyPrefab => enemyPrefab;
     public float EnemyHP => enemyHP;
     public float EnemyMoveSpeed => enemyMoveSpeed;
-    public float EnemyDamage => enemyDamage;
+    public int EnemyDamage => enemyDamage;
     public EnemyPhysicalResistance PhysicalResistance => physicalResistance;
     public EnemyMagicResistance MagicResistance => magicResistance;
 
     [SerializeField] private GameObject enemyPrefab;
     [SerializeField] private float enemyHP;
     [SerializeField] private float enemyMoveSpeed;
-    [SerializeField] private float enemyDamage = 1f;
+    [SerializeField] private int enemyDamage = 1;
     [SerializeField] private EnemyPhysicalResistance physicalResistance;
     [SerializeField] private EnemyMagicResistance magicResistance;
